requirements.txt
fastapi
uvicorn[standard]
python-dotenv
kiteconnect
python-multipart
pydantic
pydantic-settings
fyers-apiv3
pytest

sqlalchemy
pyodbc
